Moved tile info to it's own file
